Adding early-stage documentation.
This commit creates a "docs" subdirectory of the project root
in which to place relevant documentation; in general, sets of
documents should go in subdirectories of "docs", with a "NOTES.md"
file included describing what is in each collection.

This commit also adds early-stage documentation produced by
Mike Simpson during his work in July 2018; it's included mainly
for historical interest as we move forward with the Ansible
implementation.