Adding Splunk Forwarder configuration and establishing baseline practice. (#6)
This commit adds installation and configuration of the Splunk Forwarder
software to the "_common" role. Work on this configuration was done by
Isaiah Nolan.

In addition, this commit lays the foundation for the "baseline" practice
of virtual machine management and deployment; for details, see the
"Deployment Regime" document in the repository wiki. In brief, we now
maintain sequential "baseline" templates in vSphere (representing
application of the "_common" role to a vanilla OS install); all future
service deployments should start from specific baseline VM templates.

This practice keeps all of our baseline configuration explicit in the
"_common" role in Ansible, but also provides ready-made template machines
with most of the "_common" customizations already applied to be used
to create the various service nodes.