Adding SSH user keys management to the _common role.
Added by Elia Nazarenko

This introduces the ability to create\remove user accounts with or without passwordless sudo access and to install\revoke public keys from our server inventory.
The user-ssh-keys.yml task within the _common role runs in a loop through all the users located in the shell_users list in the group_vars/all/vars.yml file.
In this list we are able to set parameters for each user including username, whether to grant sudo, which shell they should use, if the user account should be created or deleted,
if the key should be installed or revoked, and the reference to the public key living in vault at group_vars/all/vault.yml

The one known gotcha in this task is that an error will be generated and the playbook will fail if user_state is set to absent but the key_state is set to present.
This condition is impossible and will create an error.

To delete a user account, first a playbook run has to be done with user_state present and key_state absent, to revoke the key.
Then another run with both key_state and user_state set to absent, to delete the user account.
This is because the user has to exist in the OS before any sort of key management can be done.

...copy and pasted my User Key playbook from ansible-ops1

...Added shell_users lists, edited user-ssh-keys to use lists and loops

...Added Public Keys to vault, vars edit, fixed variable in playbook.

..Testing Playbook, more works still needs to be done

...More changes from testing, logic for /etc/sudoer task still broken

Fixed conditional issue with sudoers task

...clean up and final touches.  Works to install and revoke public keys.

...Added more developer account public keys

...added more developer keys