Third pass with Open Arizona service management. (#8)
This commit significantly re-works the "openaz" role, bringing in
enhancements to the "_common" role and the concept of baselined
template hosts; it reorganizes the locations of a number of variables
to more accurately reflect their nature (general vs. service-specific
vs. host-specific); finally, it updates Manifold service component to
v2.0.2.J Secondarily, it reorganizes some of the work done on the
previous commit for user account and key management.

Work remaining to be done for a fourth Open Arizona pass:

*   Create an Ubuntu 18.x baseline host, and use that for the next
    deployment of the Open Arizona service nodes.

*   Convert Open Arizona to run as HTTPS rather than HTTP.

Subsequent passes should also be expected to pick up security
enhancements to the common role, as well as any point-releases of the
Manifold software.