Added email configuration settings to main Manifold configuration file.
Email settings were added based on work done by Will Simpson, see:

    https://redmine.library.arizona.edu/issues/10827

Note that the settings are going in under the "manifold_api['env']"
data structure, rather than in the more specific keys under
"manifold_api['manifold_email_*']" because the more specific keys
don't seem to be recognized, at least as of v2.0.4.

Parameterization for the new configuration values is provided
in the variables file for the "openaz" group.