Major re-write of common configuration and addition of Splunk services. (#15)
This commit completely re-works the previous "_common" role, splitting it
into more granular "_<roles>" and setting up the framework for beginning
CIS hardening of our service nodes. In general, the new version should
allow for more targeted playbook runs, e.g. "I only want to synchronize
time for a few specific hosts, not apply the entire CIS hardening suite
to everything in the inventory." This commit doesn't get us all the way
there, but it sets up the foundation on which we can build in that
direction.

This commit also adds Splunk Universal Forwarder installation and update
roles per Isaiah Nolan's original work in other branches. These roles
can be invoked by a separate playbook ("site_splunkf.yml"), again to
try to preserve granularity in application.

A number of other small changes are rolled into this commit, including
the use of "block:" directives, as well as an attempt to standardize the
grammar and syntax across all of the configuration files, to improve
readability and maintainability. There is also now a "_scratch" inventory
file to be used during Ansible development and testing: contents of that
inventory are expected to be fairly variable and/or temporary.

Finally, for now, the Open Arizona roles should be considered deprecated;
the configurations are still in place but installation and maintenance
of the service components (e.g., everything above the OS configuration)
will now be maintained manually by the Development team.